+++ RELEASENOTES.md
@@ -1,5 +1,11 @@
 # Release Notes
 
+### 1.0.21 (November 5th, 2018)
+  - add new methods (create, join, invite, leave, defaultVideo)
+
+### 1.0.20 (October 27th, 2018)
+  - upgrade the Android SDK (better connectivity support, better ui/ux)
+
 ### 1.0.19 (October 25th, 2018)
   - upgrade media to match the VU Meter